requirements.txt
llama-parse
python-dotenv
beautifulsoup4
python-docx
pdfplumber
langchain-community
langchain-text-splitters
fastapi
uvicorn
python-multipart
chromadb
langchain-chroma
langchain-ollama
pydantic
google-generativeai
groq
langchain-google-genai
langchain-groq
pandas
openpyxl
duckduckgo-search
sqlalchemy>=2.0
alembic
psycopg2-binary
openai
duckduckgo-search==5.3.1
